refactor(appointment): 약속잡기 결과 페이지 내부 컴포넌트들 폴더 구조 변경
