+++ frontend/src/components/AppointmentResult/AppointmentResultRanking/AppointmentResultRanking.tsx
@@ -1,17 +1,17 @@
 import React, { useState, useEffect } from 'react';
 import styled from '@emotion/styled';
 
-import Box from '../common/Box/Box';
-import Avatar from '../common/Avatar/Avatar';
-import FlexContainer from '../common/FlexContainer/FlexContainer';
-import Crown from '../../assets/crown.svg';
-import { getAppointmentRecommendation } from '../../api/appointment';
+import Box from '../../common/Box/Box';
+import Avatar from '../../common/Avatar/Avatar';
+import FlexContainer from '../../common/FlexContainer/FlexContainer';
+import Crown from '../../../assets/crown.svg';
+import { getAppointmentRecommendation } from '../../../api/appointment';
 import {
   AppointmentInfoInterface,
   AppointmentRecommendationInterface
-} from '../../types/appointment';
-import { MemberInterface, GroupInterface } from '../../types/group';
-import { getGroupMembers } from '../../api/group';
+} from '../../../types/appointment';
+import { MemberInterface, GroupInterface } from '../../../types/group';
+import { getGroupMembers } from '../../../api/group';
 
 interface Props {
   groupCode: GroupInterface['code'];
